Render the scene (has fish eye artifact)

+++ QuakeTraceApp.hpp
@@ -15,5 +15,5 @@ public:
 private:
     static void setIconFromAsset(SDL_Window* window, AssetHelper::ID id);
     static void renderScene(const Scene& scene, FrameBuffer* fb);
-    static std::uint32_t renderPixel(const Scene& scene, int x, int y);
+    static std::uint32_t renderPixel(const Scene& scene, float x, float y);
 };
